pyproject.toml
[tool.poetry]
name = "tooth-detection"
version = "0.1.0"
description = ""
authors = ["Votun <cabaljer@gmail.com>"]
readme = "README.md"
packages = [{include = "tooth_detection"}]

[tool.poetry.dependencies]
python = "^3.9"
opencv-python = "^4.7.0.72"
numpy = "^1.24.2"
matplotlib = "^3.7.1"
torchvision = "^0.15.1"
ultralytics = "^8.0.86"
PyYAML = "^6.0"
requests = "^2.28.2"
beautifulsoup4 = "^4.12.2"

joblib = "^1.0.1"
pandas = "^1.2.3"
python-dateutil = "^2.8.1"
scikit-learn = "^0.24.1"
scipy = "^1.6.2"
dvc = "^3.0.0"
dvc-s3 = "^2.23.0"
lxml = "^4.9.2"
click = "^8.1.3"
dash = "^2.10.2"
dash-bootstrap-components = "^1.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
